Move geo shapes to domain specific types package.
