+++ src/main/java/org/springframework/data/redis/connection/lettuce/LettuceGeoCommands.java
@@ -34,6 +34,7 @@ import org.springframework.data.geo.GeoResults;
 import org.springframework.data.geo.Metric;
 import org.springframework.data.geo.Point;
 import org.springframework.data.redis.connection.RedisGeoCommands;
+import org.springframework.data.redis.domain.geo.GeoShape;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
